refactor: lifecycle.ts -> lifecycles.ts
